gzc-core: separate location with indentation from location without indentation
