+++ lib/gzc-core/gzc-core.cabal
@@ -10,6 +10,7 @@ build-type:     Simple
 
 library
   exposed-modules:
+      Data.IndentLocated
       Data.Located
       Language.Zilch.Core.AbstractSyntaxTree
       Language.Zilch.Core.ConcreteSyntaxTree
