+++ lib/gzc-core/src/Language/Zilch/Core/Tokens.hs
@@ -1,6 +1,6 @@
 module Language.Zilch.Core.Tokens where
 
-import Data.Located (Located)
+import Data.IndentLocated (Located)
 import Data.Text (Text)
 
 -- | A simple type alias for located tokens (tokens with an additional position information).
